Switch to more recent polling
The problem is that advice where we recommended LD because the only
success is from EP + polling.

This would change to LAB. But the sudden change suggests this should
change to "no recommendation". Example: Epping Forrest

We were cautious about including a range of outcomes from the GE, but
small changes can completely change advice for the EP elections.
We may wish to apply the full "skewed polling" approach to get all the
possible outcomes?

The quick fix is to keep our early october polling and the newer
polling, which isn't perfect but causes LD->LAB to be LD->None.